% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmeans_Optimal_K.R
\name{find_Kmeans_best_k}
\alias{find_Kmeans_best_k}
\title{find_Kmeans_best_k}
\usage{
find_Kmeans_best_k(data, max_k = 10, Method = "coude", verbose = FALSE)
}
\arguments{
\item{data}{The dataset for which K-means clustering will be performed.}

\item{max_k}{The maximum number of clusters to consider. It defaults to 10.}

\item{Method}{The method used to determine the optimal number of clusters.
Acceptable values are "coude" (elbow method), "silhouette" (silhouette method),
or "gap" (gap statistics).}

\item{verbose}{Logical. If TRUE, additional output is provided.}
}
\value{
This function does not return a value but prints the optimal number of clusters
based on the chosen method and plots the corresponding graph.
}
\description{
find_Kmeans_best_k
}
\examples{
data(iris)
find_Kmeans_best_k(iris[,-5],9,Method = "coude")
}
