% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMEANS.KNN-package.R
\docType{package}
\name{KMEANS.KNN-package}
\alias{KMEANS.KNN}
\alias{KMEANS.KNN-package}
\title{KMEANS.KNN: KMeans and KNN Clustering Package}
\description{
Implementation of Kmeans clustering algorithm and a supervised KNN (K Nearest Neighbors) learning method. It allows users to perform unsupervised clustering and supervised classification on their datasets. Additional features include data normalization, imputation of missing values, and the choice of distance metric. The package also provides functions to determine the optimal number of clusters for Kmeans and the best k-value for KNN: knn_Function(), find_Knn_best_k(), KMEANS_FUNCTION(), and find_Kmeans_best_k().
}
\author{
\strong{Maintainer}: LALLOGO Lassané \email{lassanelallogo2002@gmail.com} (\href{https://orcid.org/0009-0004-1637-3511}{ORCID})

}
\keyword{internal}
